Refactored domain structure and namespaces
