+++ source/IncomingMessages.Application/IncomingMessages.Application.csproj
@@ -10,8 +10,6 @@
     <ProjectReference Include="..\MasterData.Interfaces\MasterData.Interfaces.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Dapper" Version="2.1.66" />
-    <PackageReference Include="AdaskoTheBeAsT.Dapper.NodaTime" Version="5.0.0" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.2" />
     <PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.13.2">
       <PrivateAssets>all</PrivateAssets>
