+++ source/IncomingMessages.Domain/Messages/IIncomingMessage.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 
-namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Models.IncomingMessage;
+namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Messages;
 
 /// <summary>
 /// Represents an incoming message
