+++ source/IncomingMessages.Domain/TransactionId/ITransactionIdRepository.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Models.TransactionId;
+namespace Energinet.DataHub.EDI.IncomingMessages.Domain.TransactionId;
 
 /// <summary>
 /// Store containing transaction id for all received market activity records
