+++ source/IncomingMessages.Infrastructure/Extensions/DependencyInjection/IncomingMessagesExtensions.cs
@@ -22,17 +22,17 @@ using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.Extensions.DependencyI
 using Energinet.DataHub.EDI.DataAccess.Extensions.DependencyInjection;
 using Energinet.DataHub.EDI.IncomingMessages.Application;
 using Energinet.DataHub.EDI.IncomingMessages.Application.UseCases;
+using Energinet.DataHub.EDI.IncomingMessages.Domain.MessageId;
 using Energinet.DataHub.EDI.IncomingMessages.Domain.MessageParsers;
 using Energinet.DataHub.EDI.IncomingMessages.Domain.MessageParsers.RSM012;
 using Energinet.DataHub.EDI.IncomingMessages.Domain.MessageParsers.RSM016;
 using Energinet.DataHub.EDI.IncomingMessages.Domain.MessageParsers.RSM017;
-using Energinet.DataHub.EDI.IncomingMessages.Domain.Models.IncomingMessage;
-using Energinet.DataHub.EDI.IncomingMessages.Domain.Models.MessageId;
-using Energinet.DataHub.EDI.IncomingMessages.Domain.Models.Response;
-using Energinet.DataHub.EDI.IncomingMessages.Domain.Models.TransactionId;
+using Energinet.DataHub.EDI.IncomingMessages.Domain.Messages;
+using Energinet.DataHub.EDI.IncomingMessages.Domain.Response;
 using Energinet.DataHub.EDI.IncomingMessages.Domain.Schemas.Cim.Json;
 using Energinet.DataHub.EDI.IncomingMessages.Domain.Schemas.Cim.Xml;
 using Energinet.DataHub.EDI.IncomingMessages.Domain.Schemas.Ebix;
+using Energinet.DataHub.EDI.IncomingMessages.Domain.TransactionId;
 using Energinet.DataHub.EDI.IncomingMessages.Domain.Validation;
 using Energinet.DataHub.EDI.IncomingMessages.Infrastructure.Configuration.DataAccess;
 using Energinet.DataHub.EDI.IncomingMessages.Infrastructure.Configuration.Options;
