+++ source/IncomingMessages.Infrastructure/IncomingMessages.Infrastructure.csproj
@@ -21,6 +21,8 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="Dapper" Version="2.1.66" />
+    <PackageReference Include="AdaskoTheBeAsT.Dapper.NodaTime" Version="5.0.0" />
     <PackageReference Include="Energinet.DataHub.ProcessManager.Client" Version="2.0.0" />
     <PackageReference Include="Energinet.DataHub.ProcessManager.Orchestrations.Abstractions" Version="1.5.0" />
     <PackageReference Include="Microsoft.Extensions.Azure" Version="1.10.0" />
